+++ sdk/storage/azure-storage-file-datalake/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 12.0.0-beta.11 (Unreleased)
 - Added ability to rename files and directories across file systems.
 - Added DataLakeFileSystem.getRootDirectory methods to get the root directory in a file system.
+- Fixed a bug in the download stream where the under where a buffer would be reclaimed before consumption. This issue occurs in scenarios where the stream is captured for later consumption.
 
 ## 12.0.0-beta.10 (2020-01-15)
 This package's